Revert "Refactor login into index"
This reverts commit 082842414bf1eb6a5799a4e00c3e7427f38fdc70.
