+++ src/pages/index.js
@@ -4,52 +4,15 @@ import { Link } from "gatsby"
 import Layout from "../components/Layout"
 import Image from "../components/image"
 import SEO from "../components/seo"
-import { navigate } from "@reach/router"
-import { useState } from "react"
-import StyledFirebaseAuth from "react-firebaseui/StyledFirebaseAuth"
-import { setUser, isLoggedIn } from "../utils/auth"
-import firebase from "gatsby-plugin-firebase"
+import Login from "../components/Login"
 
-const IndexPage = () => {
-  // const [firebase, setFirebase] = useState()
-
-  // setFirebase(firebase)
-
-  console.log(firebase)
-
-  if (isLoggedIn()) {
-    navigate(`/app/profile`)
-    return null
-  }
-
-  function getUiConfig(auth) {
-    return {
-      signInFlow: "popup",
-      signInOptions: [
-        auth.GoogleAuthProvider.PROVIDER_ID,
-        // auth.EmailAuthProvider.PROVIDER_ID,
-      ],
-      // signInSuccessUrl: '/app/profile',
-      callbacks: {
-        signInSuccessWithAuthResult: result => {
-          setUser(result.user)
-          navigate("/app/profile")
-        },
-      },
-    }
-  }
-
-  return (
-    <Layout>
-      <p>Please sign-in to access to the private route:</p>
-      {firebase && (
-        <StyledFirebaseAuth
-          uiConfig={getUiConfig(firebase.auth)}
-          firebaseAuth={firebase.auth()}
-        />
-      )}
-    </Layout>
-  )
-}
+/**
+ * TODO: just make the index page the login page
+ */
+const IndexPage = () => (
+  <Layout>
+    <Login/>
+  </Layout>
+)
 
 export default IndexPage
